+++ test/lib/Elastica/Test/ResponseTest.php
@@ -131,18 +131,6 @@ class ResponseTest extends BaseTest
     /**
      * @group unit
      */
-    public function testArrayErrorMessage()
-    {
-        $response = new Response(json_encode(array(
-            'error' => array('a', 'b'),
-        )));
-
-        $this->assertEquals(json_encode(array('a', 'b')), $response->getErrorMessage());
-    }
-
-    /**
-     * @group unit
-     */
     public function testStringErrorMessage()
     {
         $response = new Response(json_encode(array(